+++ src/components/Popups/PlaylistSettings.js
@@ -2,6 +2,8 @@
 //required libraries
 import React from "react";
 import { useSelector, useDispatch } from "react-redux";
+
+//Required Variables
 import { updatePlaylist, updatePlaylists } from "../Redux/Slice_Playlists";
 import { updateCurrentPlaylist } from "../Redux/Slice_CurrentPlaylist";
 import { updateName } from "../Redux/Slice_CurrentPlaylist";
@@ -12,6 +14,7 @@ import savePlaylists from "../API/savePlaylists";
 import CurrentPlaylistItems from "./currentPlaylistItems";
 import delPlaylists from "../API/deletePlaylist";
 
+//Required Components
 
 const PlaylistSettings = () => {
     const dispatch = useDispatch();
@@ -35,16 +38,25 @@ const PlaylistSettings = () => {
 
 
 
+    
+
     return(    
         <div className="settings-grid-playlist-settings flex-column">
             <div className="settings-buttons center">
-                <h1 className="button" onClick={() => {dispatch(deletePlaylist(currentPlaylist)); delPlaylists(currentPlaylist, profile.id)}}>Delete</h1>
+                <h1 className="button" onClick={() => { 
+                    
+                    if(window.confirm("Are you sure you want to delete this?")){ 
+                        dispatch(deletePlaylist(currentPlaylist)); 
+                        delPlaylists(currentPlaylist, profile.id);
+                        dispatch(updateCurrentPlaylist(playlists[0]))
+                    }
+                     }}>Delete</h1>
                 <h1 className="button" onClick={() => {
                     dispatch(updatePlaylist(currentPlaylist));
                     savePlaylists(currentPlaylist, profile.id);
                 }}>Save</h1>
                 <h1 className="button" onClick={() => {savePlaylists(newPlaylist, profile.id).then((data) => {dispatch(createPlaylist(data))})}}>Create</h1>
-                <h1 className="button" onClick={() => {getPlaylists(profile.id).then((data) => {dispatch(updatePlaylists(data))})}}>reload</h1>
+                <h1 className="button" onClick={() => {getPlaylists(profile.id).then((data) => {dispatch(updatePlaylists(data))})}}>Reload</h1>
             </div>
             <div className="center">
                 <label style={{fontSize:"21px"}} htmlFor="playlistName">Name: </label>