+++ src/components/Popups/SearchResults.js
@@ -1,6 +1,7 @@
 //Required Libraries
 import React from "react";
 import { useSelector, useDispatch } from 'react-redux';
+//Required Components
 //Required Variables
 import { addSong } from "../Redux/Slice_CurrentPlaylist";