+++ src/changelog.md
@@ -6,6 +6,8 @@
 + Added a load feature when someone logs it, it automatically reloads their playlists.
 + Fixed Issue where creating a playlist would overwrite other playlists at times. Each playlist now will have a unique number from ALL other playlists
 + Added more CSS to better format the site.
++ Added Pop up confirmation box to confirm user wants to delete a playlist
++ Added Logout button to profile drop down
 # Version 0.0.4
 + Added Modifications to CSS of Iframes and Scroll Bars. 
 + Added Dropdown to Nav Bar for following Goal